+++ app/routes/api/v2/PostsRoute.js
@@ -1,10 +1,17 @@
 import { getBySeqNumber } from '../../../controllers/api/v1/CommentsController';
-import { show, opengraph, getByIds, leave } from '../../../controllers/api/v2/PostsController';
+import {
+  show,
+  opengraph,
+  getByIds,
+  leave,
+  getReferringPosts,
+} from '../../../controllers/api/v2/PostsController';
 import { getTranslatedBody } from '../../../controllers/api/v2/TranslationController';
 
 export default function addRoutes(app) {
   app.get('/posts/:postId', show);
   app.get('/posts/:postId/translated-body', getTranslatedBody);
+  app.get('/posts/:postId/backlinks', getReferringPosts);
   app.get('/posts-opengraph/:postId', opengraph);
   app.get('/posts/:postId/comments/:seqNumber', getBySeqNumber);
   // We use POST here because this method can accept many post IDs