+++ CHANGELOG.md
@@ -30,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A `withModifiedConfig` helper function that allows to override configuration
   in test.
 
+- A new `/v2/posts/:postId/backlinks` API method that returns feed of all posts
+  that references to the given post.
+
 ## [2.13.2] - 2023-08-03
 ### Fixed
 - "Discarding" backlinks in comments when the link in their parent post is